refactor: helper
