+++ library/helper/convert_time.go
@@ -1,4 +1,4 @@
-package time
+package helper
 
 import (
 	"fmt"
