+++ library/helper/password.go
@@ -1,4 +1,4 @@
-package password
+package helper
 
 import (
 	"log"
