--- library/helper/response/message.go
@@ -1,5 +1,0 @@
-package response
-
-const (
-	Success = "Success"
-)
