+++ library/helper/validator.go
@@ -1,4 +1,4 @@
-package validator
+package helper
 
 import (
 	"fmt"
